Updated writeDimension() to take doubles instead of ints (#5310) as i…

+++ mapfile.c
@@ -525,13 +525,13 @@ static void writeKeyword(FILE *stream, int indent, const char *name, int value, 
   va_end(argp);
 }
 
-static void writeDimension(FILE *stream, int indent, const char *name, int x, int y, char *bind_x, char *bind_y)
+static void writeDimension(FILE *stream, int indent, const char *name, double x, double y, char *bind_x, char *bind_y)
 {
   writeIndent(stream, ++indent);
   if(bind_x) msIO_fprintf(stream, "%s [%s] ", name, bind_x);
-  else msIO_fprintf(stream, "%s %d ", name, x);
+  else msIO_fprintf(stream, "%s %.15g ", name, x);
   if(bind_y) msIO_fprintf(stream, "[%s]\n", bind_y);
-  else msIO_fprintf(stream, "%d\n", y);
+  else msIO_fprintf(stream, "%.15g\n", y);
 }
 
 static void writeDoubleRange(FILE *stream, int indent, const char *name, double x, double y)
